+++ truffle/CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 ## Version 1.0.0 RC9
 
 * Added support for setting the `ThreadGroup` and `stackSize` on truffle thread creation in `TruffleLanguage.Env.createThread`.
+* Added `Instrumenter.lookupExecutionEventNode()` to find an execution event node inserted at the node's location by an event binding.
+* Added `SourceElement.ROOT` and `StepConfig.suspendAnchors()` to tune debugger stepping.
 
 ## Version 1.0.0 RC8